+++ CHANGELOG.md
@@ -4,6 +4,37 @@ All notable changes to HlsKit will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.0
+
+### Added
+
+- New HLS AES-128 encryption support
+- Custom audio codec support
+- Custom audio bitrate support
+
+### Changed
+
+- Refactored Ffmpeg codebase with a new command builder to dynamically build Ffmpeg commands based on the provided settings.
+- Added encryption settings to Ffmpeg command builder, now supports AES-128 encryption.
+- Added support for custom video codec and bitrate settings within the `HlsVideoProcessingSettings` struct.
+- Started codebase improvements to decouple Ffmpeg and open to new backends such as GStreamer.
+
+### Deprecated
+
+- N/A
+
+### Removed
+
+- N/A
+
+### Fixed
+
+- N/A
+
+### Security
+
+- N/A
+
 ## 1.0.3
 
 ### Added